+++ E-Commerce/src/main/java/com/uade/tpo/E_Commerce/entity/Roles.java
@@ -15,7 +15,7 @@ public class Roles {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_role;
+    private Long id_role;
 
     @Column(name = "role_name", length = 100)
     private String roleName;
